phishing_detector/requirements.txt
# Core Framework
langchain
langchain-community
langchain-core

# LLM con capacidades multimodales
langchain-openai

# Herramientas básicas
python-dotenv>=0.19.0
requests
beautifulsoup4

# Opcional para búsqueda web
tavily-python

# Nuevas dependencias para fuentes dinámicas y análisis de logos
aiohttp>=3.8.0
Pillow>=10.0.0  # Para manejo de imágenes
imagehash>=4.3.0

# Opcional para análisis avanzado de imágenes
# opencv-python>=4.5.0  # Descomentar si necesitas análisis geométrico avanzado
# numpy>=1.21.0         # Requerido por opencv 

# Nuevas dependencias para detección de tipos MIME
email-validator>=2.0.0  # Para validación básica de emails
python-magic>=0.4.27  # Para detección de tipos MIME
langdetect>=1.0.9  # Para detección de idiomas
aiofiles>=23.1.0  # Para operaciones async de archivos

# Nuevas dependencias para optimización de imágenes
openai>=1.0.0

# API y Frontend
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6  # Para manejo de archivos
streamlit>=1.28.0
requests>=2.31.0

# Nuevas dependencias para Streamlit avanzado
plotly>=5.17.0  # Para gráficos interactivos
pandas>=2.0.0  # Para manejo de datos

# Nuevas dependencias para threat intelligence
aiohttp>=3.8.0  # Ya incluido arriba, para requests async
python-dateutil>=2.8.0  # Para manejo de fechas en threat intel

# Nuevas dependencias para análisis de attachments
python-magic>=0.4.27  # Ya incluido arriba, para detección MIME
py7zr>=0.20.0  # Para análisis de archivos 7z
rarfile>=4.0  # Para análisis de archivos RAR (requiere unrar)

# Nuevas dependencias para búsqueda web
aiohttp>=3.8.0  # Ya incluido arriba, para búsquedas async
urllib3>=2.0.0  # Para manejo de URLs 